ENH: Added Debian CPack generator STYLE: Enabled -Wall and -Wextra
Added SYSTEM flag to include_directories() calls to stop warnings
being generated for Args/Cereal/Eigen/Ceres. Currently don't know
the equivalent for ITK, so have not added -Wpedantic as that
generates a lot of warnings from ITK macros. Enabling warnings
threw up a LOT of signed/unsigned warnings, several unused variable
warnings and one uninitialized variable.

+++ Source/Core/Args.h
@@ -79,11 +79,11 @@ TRegion RegionArg(const std::string &a) {
     std::string el;
     typename TRegion::IndexType start;
     typename TRegion::SizeType size;
-    for (int i = 0; i < TRegion::ImageDimension; i++) {
+    for (size_t i = 0; i < TRegion::ImageDimension; i++) {
         std::getline(iss, el, ',');
         start[i] = std::stoi(el);
     }
-    for (int i = 0; i < TRegion::ImageDimension; i++) {
+    for (size_t i = 0; i < TRegion::ImageDimension; i++) {
         std::getline(iss, el, ',');
         size[i] = std::stoi(el);
     }
