+++ Source/Core/Masking.cpp
@@ -48,7 +48,7 @@ VolumeI::Pointer OtsuMask(const VolumeF::Pointer &img) {
     return mask;
 }
 
-std::vector<float> FindLabels(const QI::VolumeI::Pointer &mask, const int size_threshold, const int to_keep, QI::VolumeI::Pointer &labels) {
+std::vector<float> FindLabels(const QI::VolumeI::Pointer &mask, const unsigned long size_threshold, const size_t to_keep, QI::VolumeI::Pointer &labels) {
     auto CC = itk::ConnectedComponentImageFilter<QI::VolumeI, QI::VolumeI>::New();
     auto relabel = itk::RelabelComponentImageFilter<QI::VolumeI, QI::VolumeI>::New();
     CC->SetInput(mask);
@@ -57,7 +57,7 @@ std::vector<float> FindLabels(const QI::VolumeI::Pointer &mask, const int size_t
     // Relabel sorts on size by default, so now work out how many make the size threshold
     auto label_sizes = relabel->GetSizeOfObjectsInPixels();
     std::vector<float> kept_sizes;
-    for (int i = 0; i < to_keep && i < label_sizes.size(); i++) {
+    for (size_t i = 0; i < to_keep && i < label_sizes.size(); i++) {
         if (label_sizes[i] < size_threshold) {
             break;
         }
